+++ api_tests/src/payload.ts
@@ -1,19 +1,58 @@
-/**
- * cnd's REST API swap response for the split protocols (han, herc20, halight).
- *
- * Defined here until it is stabilized. Once stabilized, it can be moved to the SDK.
- */
-
 import {
     Action,
     EmbeddedRepresentationSubEntity,
     Entity,
     Link,
 } from "comit-sdk/dist/src/cnd/siren";
+import { Peer } from "comit-sdk";
+
+/**
+ * Payloads for the `/swaps/` REST API.
+ */
+
+/*
+ * The payload for POST requests to create a swap on the cnd REST API.
+ */
+interface Create<A, B> {
+    alpha: A;
+    beta: B;
+    role: "Alice" | "Bob";
+    peer: Peer;
+}
+
+export type HalbitHerc20Create = Create<HalbitCreate, Herc20Create>;
+export type Herc20HalbitCreate = Create<Herc20Create, HalbitCreate>;
+export type HbitHerc20Create = Create<HbitCreate, Herc20Create>;
+export type Herc20HbitCreate = Create<Herc20Create, HbitCreate>;
+
+// TODO: Field names in these types must match those in the SDK, should this dependency be removed?
+
+export type HalbitCreate = {
+    amount: string;
+    identity: string;
+    network: string;
+    cltv_expiry: number;
+};
+
+export type Herc20Create = {
+    amount: string;
+    identity: string;
+    contract_address: string; // TODO: This hould be token_address.
+    absolute_expiry: number;
+    chain_id: number;
+};
+
+export type HbitCreate = {
+    amount: string;
+    identity: string; // TODO: Should this be final_identity.
+    network: string;
+    absolute_expiry: number;
+};
 
 /**
  * The payload returned when fetching one swap on the `/swaps/:id` endpoint
  */
+
 export interface SwapResponse extends Entity {
     properties: Properties;
     entities: (LedgerState | LedgerParameters)[];
@@ -23,6 +62,7 @@ export interface SwapResponse extends Entity {
      */
     links: Link[];
 }
+
 /**
  * Element of the array in the payload returned when fetching all swaps on the `/swaps/` endpoint
  */
@@ -35,6 +75,7 @@ export interface SwapElementResponse extends EmbeddedRepresentationSubEntity {
      */
     links: Link[];
 }
+
 /**
  * The properties of a swap
  */
@@ -42,18 +83,19 @@ export interface Properties {
     /**
      * The status this swap is currently in.
      */
-    status: SwapStatus;
+    status: SwapStatus; // TODO: We do not return this?
     /**
      * The role in which you are participating in this swap.
      */
     role: "Alice" | "Bob";
 }
+
 /**
  * The overall status of a swap
  */
 export enum SwapStatus {
     /**
-     * The swap was created via the REST API but the communication phase hasn't yet been finalized.
+     * The swap was created but the communication phase hasn't yet been finalized.
      */
     Created = "CREATED",
     /**
@@ -71,8 +113,9 @@ export enum SwapStatus {
     /**
      * An unexpected internal failure aborted the swap.
      */
-    InternalFailure = "INTERNAL_FAILURE",
+    InternalFailure = "INTERNAL_FAILURE", // TODO: We don't use this?
 }
+
 /**
  * The parameters of a given ledger
  */
@@ -89,36 +132,25 @@ export interface LedgerParameters extends EmbeddedRepresentationSubEntity {
      * Class of this sub-entity to facilitate parsing.
      */
     class: ["parameters"];
-    properties: HanEthereum | HanBitcoin | Herc20 | HalightBitcoin;
+    properties: Hbit | Herc20 | Halbit;
 }
-export interface HanEthereum {
-    protocol: "han-ethereum";
-    /**
-     * Quantity in wei
-     */
-    quantity: string;
+
+export interface Hbit {
+    protocol: "hbit";
+    quantity: string; // In Satoshi.
 }
-export interface HanBitcoin {
-    protocol: "han-bitcoin";
-    /**
-     * Quantity in satoshi
-     */
-    quantity: string;
-}
+
 export interface Herc20 {
     protocol: "herc20";
-    /**
-     * Quantity in wei
-     */
-    quantity: string;
+    quantity: string; // In Wei.
     contract_address: string;
 }
-export interface HalightBitcoin {
-    /**
-     * Quantity in satoshi
-     */
-    protocol: "halight";
+
+export interface Halbit {
+    protocol: "halbit";
+    quantity: string; // In Satoshi.
 }
+//
 /**
  * The detailed description of the ledger state.
  */
@@ -136,16 +168,11 @@ export interface LedgerState extends EmbeddedRepresentationSubEntity {
      */
     class: ["state"];
     properties: {
-        /**
-         * Ledger event
-         */
         events: LedgerEvent;
-        /**
-         * The status of the escrow (HTLC or Invoice).
-         */
         status: EscrowStatus;
     };
 }
+
 /**
  * The ledger events related to a given step.
  */
@@ -199,6 +226,7 @@ export enum EscrowStatus {
      */
     IncorrectlyFunded = "INCORRECTLY_FUNDED",
 }
+
 /**
  * The possible steps needed on each side of the swap for its execution.
  *
@@ -212,6 +240,7 @@ export enum Step {
     Redeem = "redeem",
     Refund = "refund",
 }
+
 /**
  * An action that is available for the given swap.
  */