+++ cnd/src/with_swap_types.rs
@@ -3,7 +3,7 @@ macro_rules! within_swap_context {
     ($swap_context:expr, $fn:expr) => {{
         use crate::{
             asset,
-            http_api::{halight, hbit, herc20, AliceSwap, BobSwap},
+            http_api::{halbit, hbit, herc20, AliceSwap, BobSwap},
             Protocol, Role, SwapContext,
         };
 
@@ -12,60 +12,60 @@ macro_rules! within_swap_context {
         match swap_context {
             SwapContext {
                 alpha: Protocol::Herc20,
-                beta: Protocol::Halight,
+                beta: Protocol::Halbit,
                 role: Role::Alice,
                 ..
             } => {
                 #[allow(dead_code)]
                 type ActorSwap =
-                    AliceSwap<asset::Erc20, asset::Bitcoin, herc20::Finalized, halight::Finalized>;
+                    AliceSwap<asset::Erc20, asset::Bitcoin, herc20::Finalized, halbit::Finalized>;
                 #[allow(dead_code)]
                 type AlphaParams = herc20::Params;
                 #[allow(dead_code)]
-                type BetaParams = halight::Params;
+                type BetaParams = halbit::Params;
                 $fn
             }
             SwapContext {
                 alpha: Protocol::Herc20,
-                beta: Protocol::Halight,
+                beta: Protocol::Halbit,
                 role: Role::Bob,
                 ..
             } => {
                 #[allow(dead_code)]
                 type ActorSwap =
-                    BobSwap<asset::Erc20, asset::Bitcoin, herc20::Finalized, halight::Finalized>;
+                    BobSwap<asset::Erc20, asset::Bitcoin, herc20::Finalized, halbit::Finalized>;
                 #[allow(dead_code)]
                 type AlphaParams = herc20::Params;
                 #[allow(dead_code)]
-                type BetaParams = halight::Params;
+                type BetaParams = halbit::Params;
                 $fn
             }
             SwapContext {
-                alpha: Protocol::Halight,
+                alpha: Protocol::Halbit,
                 beta: Protocol::Herc20,
                 role: Role::Alice,
                 ..
             } => {
                 #[allow(dead_code)]
                 type ActorSwap =
-                    AliceSwap<asset::Bitcoin, asset::Erc20, halight::Finalized, herc20::Finalized>;
+                    AliceSwap<asset::Bitcoin, asset::Erc20, halbit::Finalized, herc20::Finalized>;
                 #[allow(dead_code)]
-                type AlphaParams = halight::Params;
+                type AlphaParams = halbit::Params;
                 #[allow(dead_code)]
                 type BetaParams = herc20::Params;
                 $fn
             }
             SwapContext {
-                alpha: Protocol::Halight,
+                alpha: Protocol::Halbit,
                 beta: Protocol::Herc20,
                 role: Role::Bob,
                 ..
             } => {
                 #[allow(dead_code)]
                 type ActorSwap =
-                    BobSwap<asset::Bitcoin, asset::Erc20, halight::Finalized, herc20::Finalized>;
+                    BobSwap<asset::Bitcoin, asset::Erc20, halbit::Finalized, herc20::Finalized>;
                 #[allow(dead_code)]
-                type AlphaParams = halight::Params;
+                type AlphaParams = halbit::Params;
                 #[allow(dead_code)]
                 type BetaParams = herc20::Params;
                 $fn